requirements.txt
htmltools==0.2.1
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
Pillow==10.0.0
plotly==5.15.0
scikit_learn==1.2.2
seaborn==0.12.2
shiny==0.4.0
shinyswatch==0.2.4
shinywidgets==0.2.1
xgboost==1.7.6
